+++ rollup.config.js
@@ -1,3 +1,0 @@
-/* eslint-disable no-global-assign */
-require('esbuild-register')
-module.exports = require('./build/rollup.config.ts')