+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Logs
 
+## 2023.07.01 v3.0.0
+
+1. add new package `os-local`, api docs: [README.md](https://github.com/saqqdy/node-kit/tree/master/packages/os-lang#readme)
+2. band new outputs & band new build scripts
+3. set sideEffects false
+4. upgrade all packages
+
 ## 2023.04.24 v2.5.1
 
 1. some docs work